% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Acts.UBS2.dist}
\alias{Acts.UBS2.dist}
\title{Acts of the Apostles textual variation data extracted from UBS2 (\emph{Greek New Testament}, 2nd ed.).}
\format{
A distance matrix with 55 rows and 55 columns:
\describe{
\item{P74}{Witness P74}
\item{A}{Witness A}
}
}
\source{
\url{https://zenodo.org/record/6505843}
}
\usage{
Acts.UBS2.dist
}
\description{
A distance matrix specifying simple matching distances between a set of witnesses. The distance matrix was constructed by Maurice Robinson. See Timothy J. Finney, "Varieties of New Testament Text" in \emph{Digging for the Truth: Collected Essays Regarding the Byzantine Text of the Greek New Testament}, ed. Mark Billington and Peter Streitenberger (Norden: FYM, 2014) for details of how the distance matrix was constructed. Witnesses are named according to the Gregory-Aland system.
}
\keyword{datasets}
