% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_agree_to_dist.R
\name{convert_agree_to_dist}
\alias{convert_agree_to_dist}
\title{Convert percentage agreement matrix to distance matrix}
\usage{
convert_agree_to_dist(file)
}
\arguments{
\item{file}{A file name, which can be a URL.}
}
\value{
A distance matrix.
}
\description{
The input is assumed to be comma-separated vector of percentage agreements; that is, a square table with rows and columns for the same set of witnesses in the same order. The first column and row is taken to be witness names. Table elements are treated as numeric. (Each element of the diagonal should be 100.)
}
\details{
The output is the distance matrix obtained by taking the 100 complement of the percentage agreement and dividing by 100.

This uses utils::read.csv to read the input CSV file.
}
\examples{
convert_agree_to_dist("../Raw/Acts-Hyytiäinen-CBGM.csv")

}
