% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_binarize.R
\name{do_bin_col}
\alias{do_bin_col}
\title{Convert a multi-state data frame of one column to a multi-column data frame of the binary equivalents of the multi-state values.}
\usage{
do_bin_col(fr)
}
\arguments{
\item{fr}{A data frame of one column.}
}
\value{
A data frame with multiple columns.
}
\description{
The required number of columns is one plus the floor of the base two logarithm of the maximum value in the input column.
}
