% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_binarize.R
\name{do_bin_rep}
\alias{do_bin_rep}
\title{Make a binary representation of a non-negative integer}
\usage{
do_bin_rep(x, n)
}
\arguments{
\item{x}{A non-negative integer.}

\item{n}{A number of binary digits.}
}
\value{
A vector of binary digits of length n.
}
\description{
Return a binary representation of the input using a given number of binary digits.
}
