% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_bootstrap_NJ.R
\name{do_bootstrap_NJ}
\alias{do_bootstrap_NJ}
\title{Perform bootstrap test of a neighbour-joining tree estimate}
\usage{
do_bootstrap_NJ(fr, write = FALSE, fn = "output/NJ/bootstrap.png")
}
\arguments{
\item{fr}{A data frame.}
}
\value{
A plot.
}
\description{
Estimate branch probabilities for an NJ tree estimate and plot the result.
}
