% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_counts.R
\name{do_counts}
\alias{do_counts}
\title{Compute pairwise defined counts}
\usage{
do_counts(fr)
}
\arguments{
\item{fr}{A data frame.}
}
\value{
A square matrix giving the number of pairwise defined columns for every combination of two rows.
}
\description{
Compute how many columns in a data set are pairwise defined for each combination of two rows. Given a selection of two rows, a column is pairwise defined if its elements are defined for both rows (i.e. neither row has \code{NA} for that column).
}
\details{
This code incorporates suggestions by Bill Venables (see Jan 2014 archive at https://list.science.auckland.ac.nz/sympa/arc/stat-rdownunder).
}
