% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_dist.R
\name{do_dist}
\alias{do_dist}
\title{Compute distance matrix from a data frame}
\usage{
do_dist(fr, digits = 3)
}
\arguments{
\item{fr}{A data frame.}

\item{digits}{Number of digits for rounding distances (default = 3).}
}
\value{
A square matrix giving the simple matching distance between every combination of two rows.
}
\description{
Compute distances between rows in a data frame. Elements are taken to be factors so the only meaningful result of comparing the same column of two rows is whether or not their elements match. Simple matching is used to compute the distance between two rows (i.e. witnesses). The simple matching distance between two rows is the number of columns (i.e. variation sites) that differ divided by number of columns compared. (Columns where either row has NA are not compared.) Distances are rounded to three digits.
}
