% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_NJ.R
\name{do_NJ}
\alias{do_NJ}
\title{Perform neighbour-joining tree estimation}
\usage{
do_NJ(dm)
}
\arguments{
\item{dm}{A distance matrix.}
}
\value{
A neighbour-joining tree plot.
}
\description{
Perform the neighbor-joining tree estimation of Saitou and Nei on a distance matrix and plot the result.
}
