% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_NJ_bootstrap.R
\name{do_NJ_bootstrap}
\alias{do_NJ_bootstrap}
\title{Perform bootstrap test of a neighbour-joining tree}
\usage{
do_NJ_bootstrap(fr, write = FALSE, fn = "output/NJ/bootstrap.png")
}
\arguments{
\item{fr}{A data frame.}

\item{write}{logical flag: (default = FALSE) if TRUE then a plot will be written to file fn.}

\item{fn}{a connection or character string naming the file to write to.}
}
\value{
An object of class "phylo".
}
\description{
Estimate bipartition support values for an NJ tree.
}
