% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_NJ_consensus.R
\name{do_NJ_consensus}
\alias{do_NJ_consensus}
\title{Return the majority-rule consensus neighbour-joining tree}
\usage{
do_NJ_consensus(fr, write = FALSE, fn = "output/NJ/consensus.png")
}
\arguments{
\item{fr}{A data frame.}

\item{write}{logical flag: (default = FALSE) if TRUE then a plot will be written to file fn.}

\item{fn}{a connection or character string naming the file to write to.}
}
\value{
An object of class "phylo".
}
\description{
This uses bootstrap sampling of a data frame to produce the majority-rule neighbour-joining tree with support values. Function do_reduction() is used with its default settings to reduce missing data in the input data frame. (It doesn't matter if the input data frame is already reduced.) Functions do_dist() and do_NJ() are used to produce 100 NJ trees from the reduced data frame. Support values are obtained from these trees.
}
