% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nmf.R
\name{do_nmf}
\alias{do_nmf}
\title{Perform non-negative matrix factorization (NMF)}
\usage{
do_nmf(
  fr,
  rank = 2,
  method = "ls-nmf",
  write = FALSE,
  fn = "output/NMF/output.png"
)
}
\arguments{
\item{fr}{A data frame.}

\item{rank}{NMF rank (default = 2)}

\item{method}{NMF method (default = "ls-nmf")}

\item{write}{logical flag: (default = FALSE) if TRUE then a plot will be written to file fn.}

\item{fn}{a connection or character string naming the file to write to.}
}
\value{
A basis map.
}
\description{
Do NMF to factorize an n x p matrix \emph{X} into two non-negative matrices: the n x r \emph{basis matrix W} and the r x p \emph{mixture coefficient matrix H} such that \emph{X ≈ WH}. The factorization rank r is typically much less than the minimum of n and p. The columns of \emph{W} are r factors of \emph{X}.
}
