% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_PAM.R
\name{do_PAM}
\alias{do_PAM}
\title{Do PAM (partitioning around medoids) analysis.}
\usage{
do_PAM(
  dm,
  write = FALSE,
  fn = "../PAM/Mark.UBS4.txt",
  ks = c(2, 3, 4, 5, 8, 15)
)
}
\arguments{
\item{dm}{A distance matrix.}

\item{write}{logical flag: (default = FALSE) if TRUE then write a table of partitions to file fn.}

\item{fn}{a connection or character string naming the file to write to.}

\item{ks}{A vector of numbers of clusters to be written.}
}
\value{
An object of class "pam" representing the clustering for the optimal value of k.
}
\description{
According to the cluster::pam() documentation, 'The pam-algorithm is based on the search for k representative objects or medoids among the observations of the dataset. These observations should represent the structure of the data. After finding a set of k medoids, k clusters are constructed by assigning each observation to the nearest medoid. The goal is to find k representative objects which minimize the sum of the dissimilarities of the observations to their closest representative object.'
}
\details{
The output clustering is the optimal one, namely that which produces the maximum average silhouette width. A file can be written to give a table of partitions corresponding to various values of k, as specified by argument ks. Each line of the table gives (1) k, (2) clusters enclosed in braces with medoids enclosed in parentheses, (3) witnesses that are poorly classified (as indicated by a negative silhouette width).
}
