% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_rank.R
\name{do_rank}
\alias{do_rank}
\title{For each witness, rank other witnesses by distance}
\usage{
do_rank(
  fr,
  n = 15,
  keep = "",
  alpha = 0.05,
  write = FALSE,
  fn = "output/rank/output.txt"
)
}
\arguments{
\item{fr}{a data frame.}

\item{n}{minimum acceptable number of pairwise defined columns in the reduced data frame. Values can range from 1 to the number of columns. (Default = 15.)}

\item{keep}{a row to be kept from elimination provided that it has enough defined (i.e. non-NA) elements to satisfy the minimum acceptable number condition.}

\item{alpha}{alpha value (i.e. probability of a type I error) (default = 0.05).}

\item{write}{logical flag: (default = FALSE) if TRUE then ranked distances will be written to file fn.}

\item{fn}{a connection or character string naming the file to write to.}
}
\value{
A matrix giving ranked distances for each witness.
}
\description{
This uses do_reduction() to reduce missing data then do_dist() to obtain the corresponding distance matrix. (It doesn't matter if the input data frame is already reduced.)
}
\details{
Function stats::qbinom() is used to estimate lower and upper critical limits distance (LCD, UCD). Distances that are less than the LCD or greater than the UCD are marked with an asterisk in the output.
}
