% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_trim.R
\name{do_trim}
\alias{do_trim}
\title{Drop witnesses from the input distance matrix if they are within a specified distance of a reference witness.}
\usage{
do_trim(dm, ref = "Byz", d = 0.25)
}
\arguments{
\item{dm}{A distance matrix.}

\item{ref}{the reference witness (e.g. "Byz").}

\item{d}{the specified distance (e.g. 0.25).}
}
\value{
A matrix
}
\description{
Drop witnesses from the input distance matrix if they are within a specified distance of a reference witness.
}
