% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mark.UBS4}
\alias{Mark.UBS4}
\title{Gospel of Mark textual variation data extracted from UBS4 (\emph{Greek New Testament}, 4th rev. ed., Stuttgart: United Bible Societies, 1993).}
\format{
A data frame with 267 rows and 142 columns:
\describe{
\item{Mk.1.1}{Codes representing readings found in the apparatus entry at Mark 1.1.}
...
}
}
\source{
\url{https://zenodo.org/record/4064629}
}
\usage{
Mark.UBS4
}
\description{
A data frame that specifies the readings of witnesses for variation sites found in the UBS4 apparatus. Numerals encode readings according to the order given in the apparatus: i.e. 1 = first reading, 2 = second reading, etc. There is one row per witness and one column per variation site contained in the apparatus. This data frame was constructed by Richard Mallett.
}
\keyword{datasets}
