% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_frame.R
\name{read_data_frame}
\alias{read_data_frame}
\title{Read a data frame}
\usage{
read_data_frame(file)
}
\arguments{
\item{file}{A file name, which can be a URL.}
}
\value{
A data frame.
}
\description{
Make a data frame from a comma-separated vector (CSV) file. The file is read as a table with rows for witnesses and columns for variation sites. The first column is taken to be witness names and the first row to be variation site identifiers. Data is assumed to have class "character". (Numerals are therefore treated as characters, not numbers.)
}
\details{
This is a wrapper for utils::read.csv.
}
\examples{
x <- read_data_frame("https://zenodo.org/record/6466262/files/Mark.UBS4.csv")

}
