% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_frame_CBGM.R
\name{read_data_frame_CBGM}
\alias{read_data_frame_CBGM}
\title{Read a CBGM file}
\usage{
read_data_frame_CBGM(fn)
}
\arguments{
\item{fn}{A file name, which can be a URL.}
}
\value{
A data frame.
}
\description{
Read a CBGM data file to produce a data frame of witnesses (as rows),
variation sites (as columns), and reading codes (as cells).
}
\details{
The input data file must have the following format:\tabular{ll}{
   Field nos \tab Description \cr
   1 \tab record no \cr
   2-10 \tab site addresses (e.g. 20101010) \cr
   11 \tab reading code (e.g. a) \cr
   12 \tab ? \cr
   13 \tab reading (e.g. χριστου) \cr
   14 \tab witness code (Nestle-Aland; e.g. 043) \cr
   15 \tab witness code (INTF; e.g. 200430) \cr
   16 \tab ? \cr
}


Site addresses (fields 2-10) are comprised of:
\itemize{
\item start address as four fields (book:chapter:verse:word)
\item end address as three fields (chapter:verse:word)
\item combined start address field (e.g. 20101010)
\item combined end address field (e.g. 20101010)
}

All but the following input fields are dropped:
\itemize{
\item 9 (combined start address)
\item 10 (combined end address)
\item 11 (reading code)
\item 14 (Nestle-Aland witness code)
\item 15 (INTF witness code)
}

The output is a data frame with witnesses as rows and variation sites as
columns. Reading codes "zz", "zw", "zu" are coded as NA = not available.
Column headings are variation site labels based on the combined start
address and an integer for each corresponding end address. For example, the
combined start address for the first variation site in Mark is "20101010"
and there is one corresponding end address. The combined start address is
converted to "Mk.1.1.10" (i.e. book.chapter.verse.word). Finally, a ".1" for
the corresponding end address is appended to produce "Mk.1.1.10.1" as the
label.
}
