% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\title{Recursively reduce a data frame that contains missing data (NAs) so that the number of columns where any two rows can be compared (i.e. neither has NA) is at least equal to a minimum value. Reduction is achieved by eliminating rows that cause a violation of the minimum condition.}
\usage{
reduce(df, min = 16, retain = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{min}{Minimum number of mutually defined places.}

\item{retain}{Name of a row that must be retained.}
}
\value{
A data frame.
}
\description{
Recursively reduce a data frame that contains missing data (NAs) so that the number of columns where any two rows can be compared (i.e. neither has NA) is at least equal to a minimum value. Reduction is achieved by eliminating rows that cause a violation of the minimum condition.
}
