% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_binarize.R
\name{do_binarize}
\alias{do_binarize}
\title{Convert multistate data frame to a binary data frame}
\usage{
do_binarize(fr)
}
\arguments{
\item{fr}{A multistate data frame.}
}
\value{
A data frame of binary states.
}
\description{
Convert multistate columns of a dataframe to binary columns. For example, if a column has seven states represented by seven codes (e.g. 1, 2, 3, 4, 5, 6, 7 or a, b, c, d, e, f, g) then this column will be converted to three columns with states represented by their binary equivalents distributed across the columns (e.g. 0 -> 0 0 0; 1 -> 0 0 1; 2 -> 0 1 0; 3 -> 0 1 1; 4 -> 1 0 0; 5 -> 1 0 1; 6 -> 1 1 1; 7 -> 1 1 1). Cells of the input dataframe that are not available (i.e. marked as NA) are converted to the appropriate number of NAs in the output dataframe.
}
