% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_CMDS.R
\name{do_CMDS}
\alias{do_CMDS}
\title{Perform classical multidimensional scaling}
\usage{
do_CMDS(dm)
}
\arguments{
\item{dm}{A distance matrix.}
}
\value{
A classical multidimensional scaling plot.
}
\description{
Perform classical multidimensional scaling (CMDS) on a distance matrix and plot the result. CMDS computes a set of points such that the distances between them approximate those of the distance matrix. The maximum number of dimensions of the points is set to three. The fraction of variance explained by the model is calculated and printed as the R-squared value.
}
