% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_counts.R
\name{do_counts}
\alias{do_counts}
\title{Compute counts of comparable columns}
\usage{
do_counts(frame)
}
\arguments{
\item{frame}{A data frame.}
}
\value{
A matrix.
}
\description{
For each combination of two rows, compute how many columns of a data frame are comparable . (Given a selection of two rows, a column is comparable if neither row has \code{NA}.) Returns a square matrix giving the count for every combination of two rows.
}
