% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_counts.R
\name{do_counts}
\alias{do_counts}
\title{Compute pairwise defined counts}
\usage{
do_counts(fr)
}
\arguments{
\item{fr}{A data frame.}
}
\value{
A matrix.
}
\description{
Compute how many columns in a data set are pairwise defined for each combination of two rows. A column is pairwise defined if its elements are defined for both rows (that is, neither row has \code{NA}). Returns a square matrix giving the count for every combination of two rows.
}
