% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DC.R
\name{do_DC}
\alias{do_DC}
\title{Perform divisive clustering analysis}
\usage{
do_DC(dm)
}
\arguments{
\item{dm}{A distance matrix.}
}
\value{
A divisive hierarchical clustering plot.
}
\description{
Perform divisive hierarchical clustering analysis on a distance matrix and plot the result.
}
