% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_NJ.R
\name{do_NJ}
\alias{do_NJ}
\title{Perform neighbour-joining tree estimation}
\usage{
do_NJ(dm, write = FALSE, fn = "output/NJ/output.png")
}
\arguments{
\item{dm}{A distance matrix.}

\item{write}{logical flag: (default = FALSE) if TRUE then a plot will be written to file fn.}

\item{fn}{a connection or character string naming the file to write to.}
}
\value{
An object of class "phylo".
}
\description{
Perform the neighbor-joining tree estimation of Saitou and Nei on a distance matrix and plot the result.
}
