% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nmf.R
\name{do_nmf}
\alias{do_nmf}
\title{Perform non-negative matrix factorization}
\usage{
do_nmf(
  fr,
  rank = 2,
  method = "ls-nmf",
  write = FALSE,
  fn = "output/NMF/output.png"
)
}
\arguments{
\item{fr}{A data frame.}

\item{rank}{NMF rank (default = 2)}

\item{method}{NMF method (default = "ls-nmf")}

\item{write}{logical flag: (default = FALSE) if TRUE then a plot will be written to file fn.}

\item{fn}{a connection or character string naming the file to write to.}
}
\value{
A basis map.
}
\description{
Do NMF.
}
