% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_reduction.R
\name{do_reduction}
\alias{do_reduction}
\title{Reduce missing data}
\usage{
do_reduction(frame, min = 15, keep = NULL)
}
\arguments{
\item{frame}{A data frame.}

\item{min}{Minimum number of comparable elements.}

\item{keep}{Name of a row to be preserved.}
}
\value{
A data frame.
}
\description{
Remove rows from a data frame so that the number of columns where any two rows are comparable is at least equal to a minimum value. (Here, two rows are comparable at a column if neither has missing data (\code{NA}) there.) Reduction is achieved by recursively eliminating rows that cause a violation of the minimum condition. At each step the two rows with the least number of comparable columns are identified then the least well-defined of the two is eliminated. (The least well-defined is the one with the most missing data.) One row may be marked to be kept from elimination.
}
